Download CTD data from the Bedford Basin Mooring Project

dod.ctd.bbmp can retrieve both index files and data files from the Bedford Basin Mooring Project (BBMP). Since the naming convention of the data files may alter from year to year, it is important to start by downloading an index first, to ascertain the name of a file of interest; see ‘Examples’.

Usage

dod.ctd.bbmp(
  year,
  index = FALSE,
  ID = NULL,
  direction = "DN",
  file = NULL,
  destdir = ".",
  age = 0,
  quiet = FALSE,
  debug = 0
)

Arguments

year

a numeric or character value specifying the year of interest. If this is not provided, it defaults to the current year.

index

logical value indicating whether to download an index, rather than a file. See Examples.

ID

an integer value giving the sequence number of the CTD cast in the given year.

direction

a character value indicating the direction of the cast, either "DN" (the default) or "UP".

file

character value giving the name to be used for the downloaded file. If this is NULL (which is the default) then the filename is as on the remote data server.

destdir

a character value indicating the directory in which to store downloaded files.

age

a numerical value indicating a time interval, in days. If the file to be downloaded from the server already exists locally, and was created less than age days in the past, it will not be downloaded again. Setting age=0 forces a download, so that existing files will always be updated. By contrast, setting age to a negative number prevents the updating of files that already exist locally, regardless of their age.

quiet

a logical value, passed to curl::curl_download(), which does the downloading. If this is TRUE (the default), then the work is done quietly. Until version version 0.1.12, this parameter was called silent, but it was renamed to match the corresponding argument in curl::curl_download(), which may be more familiar to users.

debug

an integer value indicating the level of debugging. If this exceeds 0, then some debugging messages will be printed. This value is passed down to related functions, but with 1 subtracted for each pass.

Value

dod.ctd.bbmp returns one of two things: (a) if index is TRUE, it returns a data frame with columns named year, id and direction or (b) a character value naming the file that was retrieved. See the ‘Examples’ section for an example of a typical workflow.

Historical Note

Sometime in prior to April 2025, the DFO server and the format of the data files both changed, and this necessitated changes to dod.ctd.bbmp(). Note that the server no longer supplies index files, so the present code tries to discover information by scraping the website ... making it very sensitive to changes and thus brittle.

See also

Author

Dan Kelley

Examples

# Download and study first BBMP CTD file of year 2024
if (interactive()) { # sidestep a pkgdown::build_site() error
    library(dod)
    # Note: cannot specify year=2025 because the URL is differently constructed
    index <- dod.ctd("BBMP", year = "2024", index = TRUE)
    destdir <- tempdir() # get temporary storage
    ctdFile <- dod.ctd("BBMP", year = index$year[1], ID = index$id[1], destdir = destdir)
    # Use oce to read, summarize and plot the data.
    library(oce)
    ctd <- read.netcdf(ctdFile) |>
        rename() |>
        as.ctd()
    summary(ctd)
    plot(ctd)
    unlink(destdir, recursive = TRUE) # clean up temporary storage
}