Download CTD data from the Ice-Tethered Profiler program

dod.ctd.itp() downloads CTD data from the Woods Hole Oceanographic Institutions Ice-Thethered Profile program https://www2.whoi.edu/site/itp/. Note that this server does not provide a searchable index, so users will need to visit the website https://www2.whoi.edu/site/itp/data/ to determine IDs of interest.

Usage

dod.ctd.itp(
  ID,
  info = FALSE,
  file = NULL,
  destdir = ".",
  age = 0,
  quiet = FALSE,
  debug = 0
)

Arguments

ID

an integer specifying the ID of the profiler that is of interest.

info

a logical value. If info is FALSE, which is the default, then dod.ctd.itp() downloads the zip-format file that can be expanded into a directory that holds a data file for each profile made by the profiler with the given ID. The name of the zip file is returned. (This file must be unzipped to access the data; unzip() is a convenient way to do this.) On the other hand, if info is TRUE, then nothing is downloaded, and dod.ctd.itp() returns a URL holding information about the profiler.

file

character value giving the name to be used for the downloaded file. If file is NULL, which is the default, then the filename on the server is used.

destdir

a character value indicating the directory in which to store downloaded files.

age

a numerical value indicating a time interval, in days. If the file to be downloaded from the server already exists locally, and was created less than age days in the past, it will not be downloaded again. Setting age=0 forces a download, so that existing files will always be updated. By contrast, setting age to a negative number prevents the updating of files that already exist locally, regardless of their age.

quiet

a logical value, passed to curl::curl_download(), which does the downloading. If this is TRUE (the default), then the work is done quietly. Until version version 0.1.12, this parameter was called silent, but it was renamed to match the corresponding argument in curl::curl_download(), which may be more familiar to users.

debug

an integer value indicating the level of debugging. If this exceeds 0, then some debugging messages will be printed. This value is passed down to related functions, but with 1 subtracted for each pass.

Value

dod.ctd.itp() returns the full pathname of the downloaded zip file, or a URL of an information site, depending on the value of info.

See also

Author

Dan Kelley