Get Index of Canadian Meteorological Stations

Look up information on a Canadian meteorological station, by a fuzzy search on information stored on a gc.ca website. The main use is to find a station ID, to be used with dod.met().

Usage

dod.met.index(
  name,
  max.distance = 0.1,
  url = "https://dd.weather.gc.ca/climate/observations/climate_station_list.csv",
  debug = 0
)

Arguments

name

character value use in a name search. The search is done using agrep() with the supplied max.distance value and with ignore.case set to TRUE.

max.distance

numerical value passed to agrep() in the station-name search.

url

the URL of the source file. The default value was tested on May 6, 2024, but this agency may change the file location, or the format, without notice.

debug

numerical value indicating the level of debugging. If this is 0, the work is done silently. For higher values, some information will be printed as the work is done.

Details

The source file is at the web location named by the url parameter. (Please contact the author, if the default value fails. It was known to work in May of 2024, but this agency might change the file location or format at any time.) The results are a data frame with information that ought to be easy to interpret. For use with dod.met(), the ID field that is of use is named "Climate.ID". See “Examples” for what might be of interest for hourly data.

Author

Dan Kelley

Examples

# Get information on the hourly meteorological datasets available
# for Halifax. Use `names(i)` to discover the other column names
# that might be of interest.
library(dod)
i <- dod.met.index("halifax")
i[, c("Station.Name", "Climate.ID", "HLY.First.Year", "HLY.Last.Year")]
#>                   Station.Name Climate.ID HLY.First.Year HLY.Last.Year
#> 328    BABINE LAKE HALIFAX BAY    1070572             NA            NA
#> 3004                   HALIFAX    8202200           1953          1963
#> 3005                   HALIFAX    8202198             NA            NA
#> 3006           HALIFAX CITADEL    8202220             NA            NA
#> 3007          HALIFAX DOCKYARD    8202240           2004          2024
#> 3008          HALIFAX KOOTENAY    8202252           2004          2024
#> 3009 HALIFAX STANFIELD INT'L A    8202249           2019          2024
#> 3010 HALIFAX STANFIELD INT'L A    8202251           2012          2024
#> 3011 HALIFAX STANFIELD INT'L A    8202250           1961          2012
#> 3012      HALIFAX WINDSOR PARK    8202255           2004          2024